build: detekt 2 (#575)
* build: detekt 2

* build: remove kover & serif report

+++ backend/src/main/kotlin/org/poweruptime/backend/features/monitor/service/MonitorDataService.kt
@@ -14,16 +14,16 @@ import org.springframework.transaction.annotation.Transactional
 @Service
 @Transactional(readOnly = true)
 class MonitorDataService {
-    fun findByIdAndType(id: ULong, type: MonitorType): MonitorData =
-        MonitorDataTable.getByType(type).let { table ->
-            table.selectAll()
-                .where {
-                    table.id eq id
-                }.limit(1)
-                .firstOrNull()
-                ?.let { table.rowToRecord(it) }
-                ?: throw BadRequestException("$type monitor data not found")
-        }
+    fun findByIdAndType(id: ULong, type: MonitorType): MonitorData = MonitorDataTable.getByType(type).let { table ->
+        table
+            .selectAll()
+            .where {
+                table.id eq id
+            }.limit(1)
+            .firstOrNull()
+            ?.let { table.rowToRecord(it) }
+            ?: throw BadRequestException("$type monitor data not found")
+    }
 
     @Transactional
     fun insert(monitor: MonitorRecord, data: MonitorData): MonitorData = MonitorDataTable
@@ -34,19 +34,17 @@ class MonitorDataService {
         }
 
     @Transactional
-    fun update(
-        oldMonitor: MonitorRecord,
-        updatedMonitor: MonitorRecord,
-        data: MonitorData
-    ): MonitorData = if (oldMonitor.type !== updatedMonitor.type) {
-        MonitorDataTable.getByType(oldMonitor.type).deleteById(oldMonitor.id)
+    fun update(oldMonitor: MonitorRecord, updatedMonitor: MonitorRecord, data: MonitorData): MonitorData =
+        if (oldMonitor.type !== updatedMonitor.type) {
+            MonitorDataTable.getByType(oldMonitor.type).deleteById(oldMonitor.id)
 
-        insert(updatedMonitor, data)
-    } else {
-        MonitorDataTable.getByType(updatedMonitor.type)
-            .update(updatedMonitor.id, data)
-            .let {
-                findByIdAndType(updatedMonitor.id, updatedMonitor.type)
-            }
-    }
+            insert(updatedMonitor, data)
+        } else {
+            MonitorDataTable
+                .getByType(updatedMonitor.type)
+                .update(updatedMonitor.id, data)
+                .let {
+                    findByIdAndType(updatedMonitor.id, updatedMonitor.type)
+                }
+        }
 }
